fix(tls,http,math): fix pre-existing compilation errors and source contract violations
- tls.context.builder.pas: fix 37 compilation errors (TBytes undefined,
  Length() on TJSONArray class, TStringArray.Text/CommaText property)
- math.transform.pas: fix Ortho overload ambiguity (Single/Double cast)
- http.static.pas: replace Set_ with SetHeader (source contract fix)
- Add phase4 plan document

+++ core/src/nextpas.core.math.transform.pas
@@ -448,7 +448,7 @@ begin
   Result := Ortho(
     ACenterX - HalfWidth, ACenterX + HalfWidth,
     ACenterY + HalfHeight, ACenterY - HalfHeight,
-    -1000.0, 1000.0);
+    Single(-1000.0), Single(1000.0));
 end;
 
 function Camera2D(const ACenterX, ACenterY, AZoom: Double;
@@ -469,7 +469,7 @@ begin
   Result := Ortho(
     ACenterX - HalfWidth, ACenterX + HalfWidth,
     ACenterY + HalfHeight, ACenterY - HalfHeight,
-    -1000.0, 1000.0);
+    Double(-1000.0), Double(1000.0));
 end;
 
 end.
